Handle postalCode format and birthDate Format Also extract hard coded String as resources Switch postalCode Int to Long

+++ app/src/main/java/com/leprincesylvain/altentest/techinaltestv2/model/user/Address.kt
@@ -3,7 +3,7 @@ package com.leprincesylvain.altentest.techinaltestv2.model.user
 data class Address(
     var city: String,
     var country: String,
-    var postalCode: Int,
+    var postalCode: Long,
     var street: String,
     var streetCode: String
 )
